# packages/api-node/openapi.yaml
openapi: 3.0.3
info:
  title: User Profile App - API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
...
  /api/posts:
    get:
      summary: List posts
      tags: [Posts]
      parameters:
        - in: query
          name: feedFor
          schema: { type: string }
      responses:
        '200': { description: OK }
    post:
      summary: Create post
      tags: [Posts]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        '201': { description: Created }
  /api/posts/{id}:
    put:
      summary: Update post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        '200': { description: Updated }
    delete:
      summary: Delete (soft) a post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /api/posts/{id}/like:
    post:
      summary: Like/Unlike post (toggle)
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/posts/{id}/comments:
    post:
      summary: Comment on a post
      tags: [Posts]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        '201': { description: Created }
    get:
      summary: List comments of a post
      tags: [Posts]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/notifications/{id}:
    get:
      summary: List notifications of a user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    /api/notifications/{userId}/read-all:
      post:
        summary: Mark all notifications as read for the user
        tags: [Notifications]
        security:
          - bearerAuth: []
        parameters:
          - in: path
            name: userId
            required: true
            schema:
              type: string
        responses:
          '200':
            description: All notifications marked as read
    /api/notifications/{userId}/clear:
      post:
        summary: Delete (clear) all notifications for the user
        tags: [Notifications]
        security:
          - bearerAuth: []
        parameters:
          - in: path
            name: userId
            required: true
            schema:
              type: string
        responses:
          '200':
            description: Notifications cleared
  /api/notifications/{id}/read:
    post:
      summary: Mark notification as read
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Marked as read }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...